DolphinQt: Add MotionPlus support to TAS input
Will manually controlling both an accelerometer and a gyroscope at the
same time be reasonable to do? No idea. Was this easy to implement
thanks to the input override system? Yes.

Fixes https://bugs.dolphin-emu.org/issues/12443.

+++ Source/Core/DolphinQt/TAS/TASCheckBox.cpp
@@ -58,7 +58,7 @@ void TASCheckBox::mousePressEvent(QMouseEvent* event)
 
 void TASCheckBox::OnUIValueChanged(int new_value)
 {
-  m_state.OnUIValueChanged(static_cast<u16>(new_value));
+  m_state.OnUIValueChanged(new_value);
 }
 
 void TASCheckBox::ApplyControllerValueChange()
