+++ Source/Core/DolphinQt/TAS/TASInputWindow.h
@@ -51,20 +51,20 @@ protected:
   TASCheckBox* CreateButton(const QString& text, std::string_view group_name,
                             std::string_view control_name, InputOverrider* overrider);
   QGroupBox* CreateStickInputs(const QString& text, std::string_view group_name,
-                               InputOverrider* overrider, u16 min_x, u16 min_y, u16 max_x,
-                               u16 max_y, Qt::Key x_shortcut_key, Qt::Key y_shortcut_key);
+                               InputOverrider* overrider, int min_x, int min_y, int max_x,
+                               int max_y, Qt::Key x_shortcut_key, Qt::Key y_shortcut_key);
   QBoxLayout* CreateSliderValuePairLayout(const QString& text, std::string_view group_name,
                                           std::string_view control_name, InputOverrider* overrider,
-                                          u16 zero, int default_, u16 min, u16 max,
+                                          int zero, int default_, int min, int max,
                                           Qt::Key shortcut_key, QWidget* shortcut_widget,
                                           std::optional<ControlState> scale = {});
   TASSpinBox* CreateSliderValuePair(std::string_view group_name, std::string_view control_name,
-                                    InputOverrider* overrider, QBoxLayout* layout, u16 zero,
-                                    int default_, u16 min, u16 max,
+                                    InputOverrider* overrider, QBoxLayout* layout, int zero,
+                                    int default_, int min, int max,
                                     QKeySequence shortcut_key_sequence, Qt::Orientation orientation,
                                     QWidget* shortcut_widget,
                                     std::optional<ControlState> scale = {});
-  TASSpinBox* CreateSliderValuePair(QBoxLayout* layout, int default_, u16 max,
+  TASSpinBox* CreateSliderValuePair(QBoxLayout* layout, int default_, int max,
                                     QKeySequence shortcut_key_sequence, Qt::Orientation orientation,
                                     QWidget* shortcut_widget);
 
@@ -75,8 +75,8 @@ protected:
 
 private:
   std::optional<ControlState> GetButton(TASCheckBox* checkbox, ControlState controller_state);
-  std::optional<ControlState> GetSpinBox(TASSpinBox* spin, u16 zero, u16 min, u16 max,
+  std::optional<ControlState> GetSpinBox(TASSpinBox* spin, int zero, int min, int max,
                                          ControlState controller_state);
-  std::optional<ControlState> GetSpinBox(TASSpinBox* spin, u16 zero, ControlState controller_state,
+  std::optional<ControlState> GetSpinBox(TASSpinBox* spin, int zero, ControlState controller_state,
                                          ControlState scale);
 };
